These fix support request #101520 (Can't log on) and bug #17359 (Brows CVS link does not include configured port numbers).
They make the following changes:

(1) CVS URLs are generated from "sys_default_domain" instead of
"sys_cvs_host".

(2) CVS URLs to repositories now end in "/", this avoids an unnecessary
redirection.

(3) In the "cvsweb" stuff, redirections to the CodeX login page no
longer specify the protocol to use.  This is in line with the way such
redirections are handled from PHP.

(4) A bug in "pre.php" is fixed that caused redirections to break if
the protocol was explicitly provided (this was the direct cause of the
problem in support request #101520).

(5) Comments were added to local.inc.dist to clarify that the cvsweb
interface always runs on the "sys_default_domain" web server, and that
if the "sys_cvs_host" machine is different, then the "sys_default_domain"
host must be able to remotely mount the cvs host's file system.