Relax the condition under which the presence of a field on the bug submission form can be configured. Required fields (fields which must be used in the bug tracking system) can now be made (in)visible on the bugr submission form. Plus a nasty bug was corrected ($field_name was used instead of $field)
It also means that these fields can initially have an empty value even if
they are not allowed to have it. This is not much of a problem since when
the bug is going to be modified the next time, the project member will be
warned about these empty values and will have to fill them in. Of course
if the fields in question shows up on the submission form the check on the
empty value will be done.

See SR #101705, SR #101707