Modified login.php in response to bug #18923.  The changes result in the system staying in SSL mode under the following circumstances:
(1) the session is secure
(2) the sys_https_host global is not empty
(3) either
(a) the sys_force_ssl global is set
(b) the sys_stay_in_ssl global is NOT set
(c) the stay_in_ssl POST variable is set

Note that the name of the sys_stay_in_ssl global is confusing. If this is 1, then the user will be presented with a choice... a better name would be "sys_stay_in_ssl_dialog"!