This commit for cvs backend feature. ( see Task #2477 )
Apply db/upgrade/db0011.sql to patch this new feature on an existing database.

The principle of this feature is to pilot from log_accum.pl the tracking and
notification of commits by commit sessions. Small modification of log_accum to register
commits and send notifications after verification in the database of the admin settings
of projects.