(This is a merge from branch CX_2_0_SUP - See Commit #4647)
Fix a scalability issue with the SQL query in charge of fetching the
field values for a given artifact. This was causing the MySQL query to
last forever when the number of generic fields used in a given tracker
was above 20. For a detailed description of the fix see bug #34218.