(This is a merge from branch CX_2_0_SUP - See Commit #5693)
Updated cross-linking mechanism.

Now, each time the pattern "xxx #nnn" is extracted from an artifact comment,
artifact description, or CVS comment, a link is automatically created to
artifact nnn where nnn is the artifact ID, and xxx should be the tracker
short name (e.g. 'bug', 'req', 'defect', 'SR', etc.).
Instead of having to use 'art #nnn", the user can type the tracker short name,
which is more convenient...

If the name typed ('xxx' above) does not match the tracker short name, a
warning is displayed. Similarly, if the tracker belongs to another project
than the one being browsed, another warning message is displayed.
These messages are displayed in order to draw the attention of the user,
but are not error messages: the artifact will be displayed in both cases.

Some projects use both the legacy trackers and the new ones. In case of
ambiguities (e.g. 'bug #123' when there is a tracker named 'bug' and the
legacy bug tracker), the system should be able to select the right page,
or will display the generic artifact if it can't choose.

Backward compatibility:
Even projects that do not use the new trackers will go through the new
mechanism. However, the behavior will be the same as before: you should
only use 'Bug #nnn' 'Task' or 'SR', unless you want to link to an
artifact belonging to another project.

Related artifacts:
- task #3148
- SR 147 on Partners
- Bug #34349

Upgrade notes:
The file SF/utils/cvs1/log_accum should be copied to /usr/local/bin (or
wherever you installed it).