(This is a merge from branch CX_2_0_SUP - See Commit #5942)
Short project description now displayed as pure HTML in the Software
Map and not as text. This is to be consistent with the Project Summary
page where the short description is displayed as HTML. Project
Registration form was also modified to accept HTML tags in the project
short description.
(See bug #276)