(This is a merge from branch CX_2_0_SUP - See Commit #7140)
Fix for SR #428
If a tracker field does not have the right type (e.g. the user typed a string where the system expected a float), an error message is now displayed, warning the user that the field was incorrect but also that the artifact was created.
In this case, the default value of the field is stored for the new artifact.