(This is a merge from branch CX_2_0_SUP - See Commit #7942)
Added way to submit HTML documentation without modification
Now, when documentation in HTML is uploaded, it is stored, and will be displayed, "as-is".
If the documentation is copy/pasted, it will be embedded in a CodeX page.

From the documentation:
"Tip: Copy-pasted vs. uploaded documents
A copy-pasted document will be embedded in a CodeX page, so its appeareance will be modified, and some tags like HTML, TITLE, or HEAD will be ignored. An uploaded document, on the opposite, will be displayed "as-is" in the browser window. Another difference is that a copy-pasted document can be edited online, whereas an uploaded document can only be updated by uploading a new version."

See SR 46 on Partners
https://partners.xrce.xerox.com/tracker/?func=detail&aid=46&atid=199&group_id=120
See also Task #437