(This is a merge from CX_2_2_SUP - See Commit #15286)
Improve performance of SQL query retrieving MY ARTIFACTS.
Cut it into two subqueries:
* one retrieving all trackers in groups the user is member of and there all
 artifacts the user has submitted or are assigned to him
we suppose that users can only be assignee in trackers where they are group
member
* the second query retrieving all trackers in groups the user is not member
of and there all the artifacts the user has submitted

For that to be correct we have to limit the scope of the patch committed in
commit #13920 (branch CX_SUP_2_2) and commit #13921 (trunk). I.e. from now
on, the assigned_to and multi_assigned_to fields are per default bound to
"artifact_technicians" and can not be bound to anything else. This is also
more coherent with the permissions a user has on a tracker. A user who is
not able to update an artifact in a tracker (only artifact_technicians can)
should not be an assignee.

=> fix bug #1396