(This is a merge from branch CX_2_2_SUP - See Commit #17328)
Added a starting dot "." to the domain name when setting cookies.
This enables subdomains (i.e. project web servers) to get the session cookie
on Mozilla based browsers.

This modification was first introduced in revision 1.112 of session.php
(May 7 2002, before the CVS tracking system was in place) and reverted
the day after for incompatibilities with Netscape 4.7.
Actually, with Netscape 4.7, the cookie is not set for subdomains, but it
is available on the main domain, so it is acceptable today.