(This is a merge from branch CX_2_4_SUP - See Commit #21940, which is itself a merge from branch CX_2_2_SUP - See Commit #21914 and commit #21939) This commit also integrates patch 201 submitted by ST on Partners: https://partners.xrce.xerox.com/tracker/?func=detail&aid=201&atid=203&group_id=120
This commit fixes SR 181 on partners. The symptom was that whatever the
changes in the configuration it was impossible to upload a document
bigger than 16MB in the document manager. The problem was coming from
the fact that the stock php RPMs of RHEL 3 are linked against mysql 3
and the MySQL 3.x client library has a known upper limit of 16 MB on
the packet size.

** HOW TO FIX THE PROBLEM - PLEASE FOLLOW THESE INSTRUCTIONS **

1) Install the new PHP RPMs

In order to correct the problem you must install a new version of php
compiled against the MySQL 4 client library which has a upper limit of
1GBytes on packet size. The php RPMs to install are located at:
https://codex.partners.xrce.xerox.com/packages-rhel3/RPMS_CodeX/php/php-4.3.2-23.ent.codex
Make sure you only upgrade those RPM that are installed on your system

2) Update your CVS working copy

** LIST OF MODIFICATIONS **
- SF/www/docman/doc_utils.php: do not select the document data if not
needed
- SF/www/docman/new.php and SF/www/docman/admin/index.php: make the upload
code compliant with PHP 4.3 guidelines
- SF/etc/php.conf add tmp directory and remove file limit on php script.
- documentation/installation_guide/html/CodeX_Installation_Guide.htm:
improve the installation guide to explain in details what the value
of the various parameters must be in local.inc, my.cnf and php.ini

3) Modify your /etc/httpd/conf.d/php.conf file according to the new version
stored in /home/httpd/SF/etc/php.conf.