(This is a merge from branch CX_2_4_SUP - See Commit #23814)
Applied patch 250 from ST on Partners.

When phpwiki sends an email (wiki monitoring), the content of the page used to be transmitted even if the receiver doesn't have the right permission on the corresponding page.
Now, the content of the diff is no longer sent but people are aware of page update.

Fix for SR 249 on Partners:
https://partners.xrce.xerox.com/tracker/?func=detail&aid=249&group_id=120&atid=199