(This is a merge from branch CX_2_4_SUP, see rev #1917)
Documentation generation no longer uses CVS synchronization.
In order to update CodeX documentation, simply do a 'svn update' in /home/httpd.
When 'generate_doc.sh' is launched (in 'sourceforge' crontab), it will compare
the dates of the XML files with the date of the PDF document.
If one XML file is newer, the documentation is regenerated.