Slightly improve performances for artifact reports, artifact submissions and artifact displays, especially for trackers with lot of fields and/or lot of artifacts.
Before this revision, each artifacts and artifact fields had the responsability to decide if the user has access to the object or not (view, submit, read, ...)
This means that you had to create an object to know if the user can access to. It is a pain, especially for reports of trackers with lot of artifacts.
This also means that we couldn't share the information, and when displaying or updating an artifact, each fields had to do a sql request for each permission.

The solution retained is a dedicated object PermissionsManager which is responsible to do two things :
1. retrieving permissions for an object_id (we do not have the need to create the object anymore)
2. caching those permissions. For exemple, instead of retrieving only the permissions for a field, we retrieve all permissions for all the fields of the tracker. The final query last a little bit longuer but we've got all the information needed for future calls.