License approval is now an option defined at the package level. By default, the license popup is displayed, but it is now possible to disable it.
File references are now working properly. Use 'file #123' to automatically link
to file ID 123. The file ID is displayed in a tooltip above the file name in the 
file listing (like for documents).
If licence approval is disabled for the package, the file is directly downloaded.
Otherwise, the user must approve the license first.