Add parameter in local.inc to specify the default trove category displayed in the Software Map welcome page. Default is: $sys_default_trove_cat=18; (Category 18 is 'topic').
This parameter is useful for CodeX servers that don't want to use the
'topic' category.

Please note that the parameter is not mandatory. If it is not set in
local.inc, the default value of 18 is used.

** Note for CodeX servers administrators **
If you wish to use this parameter, simply add it in /etc/codex/conf/local.inc.
You may use the diff in local.inc.dist as a template. See for instance:
https://partners.xrce.xerox.com/svn/viewcvs.php/dev/trunk/SF/etc/local.inc.dist?roottype=svn&r1=text&r2=text&tr1=2926&tr2=2925&root=codex&diff_format=l