(This is a merge from trunk branch /dev/trunk/, see rev #3061 for details)
Tracker export enhancement : it is now possible to export only the fields of the current report.

Before, tracker export exported all the fields of the tracker.
Now, a checkbox just around the export button let you specify that you want to export only the report fields.
If the checkbox is not checked, the export use the same behavior as before.