(This is a merge from support 2.8 branch /support/CX_2_8_SUP, see rev #3229 for details)
"Changes by: XXX " info does not appear in emails generated by a tracker,
for tracker without 'Assigned to' field. This was firstly design not to notify
people who couldn't read the 'assigned to' field.
In the case that the tracker doesn't have any assigned to field (neither single, neither multi),
the information will be displayed in the email.