This scripts fixes some issues with tracker field permissions on server that upgraded from CodeX 2.4 (yes, that was a long time ago). It also fixes permissions for CodeX 2.8 servers on the 'stage' field, which currently has no permission.
The script will simply add permissions for fields that have no permission set,
and that are used in many trackers (this is to avoid fields for which permissions
have been specifically removed by a project admin).

The default permission added are:
- anyone can read the field
- project members can update the field.

This is a follow-up to rev #3323.

Note to project admins:
In order to do the update, simply run check_field_perms.pl.
The script will not harm your server: you can run it several times without problem.