(This is a merge from dev/trunk. See rev #3340)
Fix email address returned when the user is not found.
The email address is now "email-not-found" instead of
"Error - Not Found" and the French version "Erreur - Introuvable".
So now:
- you only need to catch one email address (email-not-found) if you want to
be informed that there is an error with one user email address.
Simply add a line in your /etc/aliases file to catch such emails.
- The wrong email is now one word instead of three, which caused
problems with some mailers (in particular, the '-' email address
was confusing).

This is a fix for SR #499 and a way to close patch #500 (which was not
applied).