Implementation for SR #270 : Need explanation when browsing tracker for first time and getting my artifacts only.
When a logged in user browse the artifacts for the first time (or has no preferences defined), the former behaviour was to display only 'his/her' artifacts, which could be a little disturbing when s/he expected to see more aetifacts.
The new behaviour is to display all the artifacts.