Development for SR #149 : Enable to reset Date Default Value to void (rather than 0).
The default value for a date field is the current date, but it not precised really explicitly.
Now, when defining a default date value, the user can choose between 'current date' and a date value.
This also fix the issue of coming back to 'default value' after a 'date value'.