Bug fix for bug #12429 (on codex central) : Excel spreadsheets not supported by doc mgr?
When uploading an excel (xls) file in the document manager, the mime-type were wrong, and so the file was given to upload or open as a microsoft Word document.
This bug is similar to SR #267 (on Partners) and so, we use the same fix as the one used in rev #2375, by checking some suspect mime-type with the file extension.