(This is a merge from 2.8 support branch, see rev #3659 for details. We just add some code for mixing the bug fix and the new developments on private news)
Bug fix for SR #521 : Latest News on Codex front page.

In the Latest News section on Codex front page, when clicking on the [XML] link, the rss page displays all the news, including those which were not approved by Codex site admin to be promoted on front page.
This bug was introduced in rev #3484, fixing the bug reported in the SR #515.