Bug fix for bug #12376 on codex central server: cc field error states to use the back button but the bug has already been submitted.
In legacy bug tracker, if a user give a wrong name in the CC field, the bug is submitted anyhow, and an error message is displayed. By clicking the back arrow and correct the CC name, when the user submit again the bug, it creates another bug with the good CC.
Now fixed, the validity of the CC field is checked after inserting the bug in the database.
We check the attached files too.