Bug fix for cx3 #552: adding mandatory fields with CLI Client.
When adding an artifact with the CLI client, but forgetting a mandatory field, the returned error message was not really clear (returned by exit_error function, i.e. a whole HTML page)
Now, in the function exit_error, we test before if we come from the soap API or not, and return a soap fault (xml) if so.
If we don't come from the soap API, we return the usual message.