Fix for cx3 #553: permission handling in CLI Client.
If the user doesn't have full access to an artifact, the open and total number of artifacts are not displayed in the web interface but were displayed through the CLI Client.
Now fixed. If the user has not full access to the artifacts, the number of open and total are hidden in the CLI.