(This is a merge from CX_3_0_SUP branch, see rev #4200)
Fix for SR #517: Assigned To getting automatically changed, when user account is not active.
Now, if the user account is no more active, we don't select the first user "by default" but we select an additional value "unknown value".
This behaviour will warn the user that the value don't exist anymore, but will however let the user updating the artefact.
If the unexisting value is not changed, it will keep the same value (the value is still stored in database)