(This is a merge from CX_3_0_SUP support branch, see rev #4223)
Fix for:
SR #520: subversion does not recognize capital letters into CodeX groups, and
SR #526: Not easy readable layout for large number of members in SVN admin page.

The ugroup names are now displayed with their case in the subversion administration, in order to keep the same presentation that in the ugroup administration.
Please note that subversion doesn't care the case and will apply the permissions whatever the ugroup case is written.
Moreover, we now add a space after the comma in the group member list (in the subversion perms admin) to avoid the horizontal scrollbar (SR #526).