(This is a merge from CX_3_0_1_SUP. See rev #5518)
Allow CodeX servers to run without a DNS and subdomains.
This is useful for smaller CodeX servers, or for installing
CodeX in networks where there is no DNS delegation possible.

There is now an option in local.inc called 'sys_disable_subdomains'.
When set to '1', CodeX will not generate DNS entry, and will not
use subdomains like svn.proj.codex... or http://www... or cvs... or lists... etc.
Please note that in this case, project web sites are DISABLED.
(You should disable the homepage service in the projet template: it is 
not done automatically)

Note to server administrators:
You don't need to update your local.inc file: if the new parameter
is not set, then CodeX will run as before.
Also note that you should not set the parameter to '1' on an existing
server. If you want to switch to this mode, please contact us.