Fix for SR #766 : SOAP function that return the structure of one tracker.
The SOAP function that return  all the structure of all the trackers (getArtifactTypes) could be quite slow if the project contains lot's of trackers.
So we add a new SOAP function (getArtifactType) that return only the structure of a given tracker.
We add also a function for listing the trackers of a project, without all the structure (return ID, name, description, etc.)

The corresponding function in the CLI are:
- trackerlist (light list of trackers)
- trackers (list of trackers with structure)
- tracker (list of a specific tracker with structure)