(This is a merge from CX_3_2_SUP. See rev #6531 for details)
This is the "Passerelle CVS - Dimensions" Plugin.
This plugin exports a project's tagged CVS source code to an external Dimensions repository.
This module was developped specifically for France Telecom, and won't run easily on other environments.
See Readme.txt for more details.

Note to system administrators: The plugin should not be activated except by France Telecom.