(This is a merge from CX_3_2_SUP. See rev #6831)
Fix for SR #928 : Remove extension filtering for wiki attachments.
Wiki attachments cannot be executed on the server, so the risk is limited.
The only issue is with attachments downloaded, then executed on the client.
However, we do not limit extensions for docman or file release manager (which have the same issue), so there is no reason to keep this limit for wiki attachments.