(This is a merge from CX_3_4_SECURITY support branch, see rev #8753)
Fix for SR #1438 : ALL artifacts are mass-changed when trying to mass-change just a partial selection

The problem occured only if the partial selection was given by a filter on dates.
On all other filter criteria, the masschange worked fine.

The mass change needed the operator (<, = or >) to filter on dates, which was not passed before.
It is now fixed.