Add a notion of version to the SOAP API. This aims to check compatibility between clients and server.
**************************************************
Warning to Codendi web service client developers :

Now, the login function requires a version parameters.
A check will be done (only at login) to see if the client is compatible with the API you're pointing.
Then you need to add this parameter in your login function 
(the expected value of this param is the result of the function getAPIVersion of the used API).
**************************************************