(This is a merge of branch /contrib/st/bugfix/1585_remove_pre_forum contributed by ST) (See also rev #9997 on /dev/trunk)
Remove <pre> and use nl2br to format messages in forums. <pre> was
introduced for people who wanted to add code in their messages but it makes
text unreadable (very long lines). A solution whould be to allow some kind of
formatting in message to handle this case.

(Thanks Manuel)

level=minor