(This is a merge from /dev/trunk/. See revision #11985 for details) Contrib from Manuel Vacelet to fix SR #1857 - Tracker reports very slow after 3.6 upgrade
This is a merge from:

revision #11791
util_user_link use cached DB result (for tracker reports with user names)

revision #11883
Save the forest: STAIGHT_JOIN allow to better restrict the dataset (artifacts
are filtred first, before beeing joined to other tables).

revision #11884
Save the forest and avoid to fetch the same data hundred of time:
- The tracker report query already fetch the artifact_field_value: their is
no need to fetch it again right after. Except for one case: multiple select
boxes.
- When data are dressed-up for output, when it's about usernames, there
is no need to look for username based on user_id when the username is already
in the first query and there is no need to trigger a query when the only value
is "None" (even in multiple select box case).

level=normal