Make tree view blob sizes fallback gracefully on older versions
The tree view used -l indiscriminately to get blob sizes, without
taking into account that the -l option only appeared in git 1.5.3,
in which case the tree would fail to display anything (because the
command returned only the help message because of the invalid -l
parameter).  Now it falls back properly depending on the version,
only displaying blob sizes if running on git 1.5.3 or greater.