Fix ajax blame with non-geshi blobs
The non-geshi blob output has a row for each line, which is even harder
to inject blame data into.  It just made more sense to change the
non-geshi blob output to the same table structure as geshi, so one
javascript blame solution worked for both.

Plus, it had the advantage of fixing the bug that made geshi introduce
this format in the first place - if you highlight code with your mouse,
the line numbers are no longer selected along with it in non-geshi view.