Fix trees with multiple names getting mixed up
This addresses three issues:
1. Git stores trees as a hash of their content, which has nothing to do
with their name.  Therefore, two different directories with the exact
same content but different names could end up pointing to the exact same
tree hash.  This caused problems on the tree view, when it attempted to
look up the tree name by its hash - this is actually something you just
can't do reliably in git because of this system.
2. Previously, the tree view used the object cache when enumerating
items in the tree.  However, because of the problem above, you could
have two copies of the same tree under two different names.  Because the
object cache stores by reference, fetching and populating the tree with
data during its second appearance in the tree actually affected and
overwrote what was populated during its first appearance in the tree.
This happened during the model load, before the template was rendered,
so instead of showing the first instance and then the second instance,
it looked like the second instance appeared twice.
3. Previously the FilesystemObject class used a system where you would
assign the name to an object.  This was a leftover from before it was
split out of the Tree object.  Doing it this way meant there were
inconsistencies all over the place as far as what was being set for the
name - sometimes it was just the base file name, sometimes it was the
full file path, and every now and then it would have to be converted
back and forth.

Now, these are fixed:
1. Now, the f= parameter is taken as the source of truth and used to set
up the path for the tree/blob.
2. Now, the tree enumeration doesn't use the cache in order to prevent
this situation.
3. Now, for consistency, the ability to set a name for a filesystem
object is removed.  You can only set the full path now using SetPath,
SetName is removed.  The name can still be fetched but it is
automatically derived as the base file name of the path.