Fix tree link for heads
The tree link by heads was passing the old style h=refs/heads/[head]
piece of the url, from the old version.  It actually wasn't worth it to
implement the translation of the ref path to the tree object in the tree
controller, because we were already passing the hash of the head's HEAD
commit as the hashbase parameter (hb=) - in that case the tree
controller is already automatically looking up the tree for that
hashbase commit.