Add support for tags pointing to other tag objects
Tags now store the object they point to and the commit separately.  (For
tags pointing right to commits, the object and the commit are the same)
This keeps the object link working correctly by pointing to the direct
object (as opposed to dereferencing all the way down to the commit), so
you can use the object link to trace tagged tags all the way down to the
base commit.

Also, the project and the tag list page now use the same tag listing
template fragment to make things easier.