Fix loading image urls on some browsers
Some browsers don't properly determine the base url, when you use a
relative path in an href/src (eg "images/tooltip-loader.gif") - they
will treat slashes in a GET variable as a part of the actual path.  So
when you have a URL like
(http://server/gitphp/index.php?p=php/gitphp.git), despite the slash in
the GET variable being escaped in the url, trying to load the images
path above resolves as
http://server/gitphp/index.php?p=php/images/tooltip-loader.gif, which
isn't right.  This uses a regex in the javascript to extract only the
actual dirname of the instance, to give the loading img tags a full url.