Allow on the fly change of locale
This only adds the GET variable to do the switch, the UI option isn't
there yet.
This parsing needs to be done really early because we need to do the
language switch/load before anything else happens... however I really
don't like parsing get variables so early in the index like this because
it's breaking the controller scheme that everything else uses.  I may
not be able to get around that, but I'd at least like to clean it up and
maybe move this code out of the index into a function or something.