Allow user to change language without explicitly clicking "Set"
This uses javascript to submit the form as soon as the user changes the
language selection box, rather than having to also click "Set".  It
removes the Set button also.  Since this is done in the jQuery ready
event, there is a bit of time while the page is loading where the
onchange won't work and the set button will still be visible, then it'll
"disappear" when the page finishes loading.  There's no good way around
that though if we want the page to degrade properly without javascript.