Bug linking support
This allows specification of a regex and url to find bug references in
commit messages, and allows them to be linked to an external bug
tracker.  Currently this is just a very basic regex find and replace,
there's no advanced logic such as finding patterns that span lines, or
other preprocessing.
Currently this also only display in places where the full message text
is shown, and is not a link - the commit page, the commitdiff page, and
the log page.  In the various other places where commit messages may be
shown (shortlog, headers, summary, etc), the message text is already a
link to the commit itself.  Putting a link inside a link is weird
usability wise since that's not the way the user is expecting things to
behave.

Enhancement request #43