Make newly added blocks align correctly in side-by-side view
Added blocks of code (so all '+' lines) are aligning incorrectly - they
appear one line higher than where they actually are, which is confusing
because it generally makes the code look like it's syntactially invalid.

This change does fix it, but I'm really don't think that it's the right
solution.  Labeling it as a HACK for now to revisit later.