Do not rely on googlecode to provide html5shim.
'[html5shim] is the defacto way to enable use of HTML5 sectioning elements in legacy Internet Explorer, as well as default HTML5 styling in Internet Explorer 6 - 9, Safari 4.x (and iPhone 3.x), and Firefox 3.x.'
https://github.com/aFarkas/html5shiv

Not sure that providing it through our regular scripts is the clean way to go. Wanna build a dedicated rpm?