Display search results as a bunch of cards instead of tree grid in planning
Tree grid is technically a nightmare to handle the drag 'n drop:  
* table/tr elements are not really supported by scriptaculous' dragdrop library
* dragging a node means to drag its children at the same time (in a tree grid you have to fire the drag event on each subchild. No way.)

One solution would be to display the hierarchy of artifacts as a ul/li-based tree (semantically better than the table) but displaying the other data
in a tabular structure in css (fake columns) is not really better and/or easier.

The solution choosen here displays the artifacts as nested cards/PostIts like on a cardwall. Additionnal information (aka 'columns') is
not anymore diplayed but instead is displayed the 'PostIt semantic' which does not exists yet but will allow admin to define the 
content of a PostIt (which fields, wich color, ...). This semantic will be also used in the CardWall to have a unified rendering.
For now only the title and the link to the artifact is displayed.

Pro tip: add the class 'planning-droppable' to a card and you will be able to drag 'n drop items inside the search content view!