Do not index permissions for now
The permissions are not exported correctly. Instead of having a bancal
situation, we prefer to not index now the permissions.

Test case:
`- folder
   `- new doc

folder has the following permissions:
 - project admin = manager
new doc has the following permissions:
 - registered users = read

In elasticsearch the following is stored for 'new doc':
_source: {
  id: 55
  title: "new doc"
  description: "",
  permissions: {
    1: [ ]
  }
}

Expected result: as project admins can manage the folder and only them
can access the content of the folder, the permissions in elasticsearch
should allow them to access the new doc.