Fixed cardwall drag and drop bug
Drad-and-drop js callback was using the <select> tag to retrieve the id
of field used to generated the cardwall columns.
And the renderer didn't set the "selected" attribute on the <option>
tag.
An exception was raised in the AJAX request, but was ignored
client-side.

The <select> tag rendering was probably broken during some refactory:
- fetchCards() was extracted from fetch().
- getField() was extracted from fetchCards().
- the $used variable instanciation ended up in getField(), where it is
  not event used.
- it was not defined anymore in fetch(), where it was actually used.
- the isset() statement may even have been added later, to make some
  "undefined variable" exception disappear, without fixing the root bug.

I used the id of getField() instead of $this->field_id, since the later
is not aware of permissions... (and should probably not be used anymore)

I may have extracted some isTheRightField() method, but I didn't know
how to name it, since the "field" attribute name means nothing at all...

Wishlist:
- Document (e.g. $matching_ids)
- Extract a factory (and stop using both $this->getField() and
  $this->field_id in the same class)
- Move queries from fetchCards() to some DAO
- Write tests